updated to use whatever local host is avaliable
