--- bia-test-data/bia_test_data/test_api/__init__.py
@@ -1,2 +1,0 @@
-from pathlib import Path
-bia_test_data_dir = Path(__file__).parent.parent / "data"
