--- bia-test-data/bia_test_data/test_api/docker-compose.yaml
@@ -1,31 +1,0 @@
-services:
-  biaint-mongo:
-    image: mongo:6.0
-    restart: no
-    env_file: ../../../api/.env_compose
-    ports:
-      - 27018:27017
-    expose:
-      - 27017
-    healthcheck:
-      test: echo 'db.runCommand("ping").ok' | mongosh localhost:27017/test --quiet
-      interval: 1s
-      timeout: 1s
-      retries: 5
-  bia-integrator-api:
-    image: ghcr.io/bioimage-archive/bia-integrator-api:0.2.3
-    env_file: ../../../api/.env_compose
-    container_name: test_api
-    hostname: test_api
-    ports:
-      - 8081:8080
-    expose:
-      - 8080
-    depends_on:
-      - biaint-mongo
-    healthcheck:
-      # !! important to sync services
-      test: ["CMD", "wget", "-O", "-", "test_api:8080/openapi.json"]
-      interval: 1s
-      timeout: 10s
-      retries: 10
